Update readme.txt
The objective of this project is to build a pair of robotic eyes and create an Android app so that the eyes can be controlled by the app. Using the joystick and buttons on the app, commands are sent via a Bluetooth connection, allowing the eyeballs to be rotated and the eyelids to be opened or closed. Two popular microcontroller/microprocessor platforms, the Arduino UNO R3 and the Raspberry Pi 4, are used for control.

I also implemented a small program using the Raspberry Pi and computer vision so that the eyes can track and follow a human face. The eyeballs move in response to the position of the face in front of them, creating the illusion that the robotic eyes are actively observing and following the person.